refactor: update remaining import paths after modularity refactoring
Update all call sites that were still importing from the old module
locations: expandable_quote sentinels from providers/base, subagent
helpers from hook_events, UUID_RE from hook.py, view_window in
send_command, and unwired_save defaults on thread_router, user_preferences,
and window_state_store.
